backend/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
sqlalchemy>=2.0
pydantic>=2.0
pyyaml>=6.0
python-jose[cryptography]>=3.3
bcrypt>=4.0
